medsProject/requirements.txt
asgiref
django
pytz
setuptools
sqlparse
django-environ
django-tables2
django-bootstrap3
django-filter
python-dateutil